requirements.txt
pandas==2.2.2
openpyxl==3.1.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
xlsxwriter==3.1.2
click==8.1.7
